# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Claimbase Expense Claim Management System API
servers:
  - url: /api
    description: Base API path
security:
  - TelegramAuth: []
...
paths:
...
  /users/{userId}/companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyResponse"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
  /companies/{companyId}:
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Company deleted
...
  /users/{userId}/receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipts for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReceiptResponse"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Create a receipt
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptBody"
      responses:
        "201":
          description: Receipt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponse"
  /receipts/{receiptId}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get a receipt by ID
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponse"
    put:
      operationId: updateReceipt
      tags: [receipts]
      summary: Update a receipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReceiptBody"
      responses:
        "200":
          description: Receipt updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponse"
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete a receipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Receipt deleted
  /users/{userId}/resolved-flags:
    get:
      operationId: listResolvedFlags
      tags: [receipts]
      summary: List resolved flags for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of resolved flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResolvedFlag"
    post:
      operationId: resolveFlag
      tags: [receipts]
      summary: Resolve a flag on a receipt
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiptId, flagType]
              properties:
                receiptId:
                  type: integer
                flagType:
                  type: string
                  enum: [dupe, purpose]
      responses:
        "200":
          description: Flag resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /receipts/parse-image:
    post:
      operationId: parseReceiptImage
      tags: [receipts]
      summary: Parse receipt image using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseReceiptImageBody"
      responses:
        "200":
          description: Parsed receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedReceiptData"
  /currency/rate:
    get:
      operationId: getExchangeRate
      tags: [currency]
      summary: Get suggested exchange rate
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exchange rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponse"
...
components:
  schemas:
    ResolvedFlag:
      type: object
      properties:
        receiptId:
          type: integer
        flagType:
          type: string
      required:
        - receiptId
        - flagType
...
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        baseCurrency:
          type: string
        cutoffDay:
          type: integer
        cutoffTime:
          type: string
        cutoffMonthOffset:
          type: integer
          description: "0 = same month, 1 = following month"
          default: 1
      required:
        - name
        - baseCurrency
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        baseCurrency:
          type: string
        cutoffDay:
          type: integer
        cutoffTime:
          type: string
        cutoffMonthOffset:
          type: integer
          description: "0 = same month, 1 = following month"
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        baseCurrency:
          type: string
        cutoffDay:
          type: integer
        cutoffTime:
          type: string
        cutoffMonthOffset:
          type: integer
          description: "0 = same month, 1 = following month"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - baseCurrency
        - cutoffDay
        - cutoffTime
        - cutoffMonthOffset
...
    CreateReceiptBody:
      type: object
      properties:
        companyId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        receiptDate:
          type: string
          format: date
        currency:
          type: string
        amount:
          type: string
        conversionRate:
          type: string
        convertedAmount:
          type: string
        imageUrl:
          type: string
        claimMonth:
          type: string
          nullable: true
          description: "YYYY-MM override for which claim period this receipt belongs to"
      required:
        - companyId
        - description
        - receiptDate
        - currency
        - amount
        - conversionRate
        - convertedAmount
    UpdateReceiptBody:
      type: object
      properties:
        companyId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        receiptDate:
          type: string
          format: date
        currency:
          type: string
        amount:
          type: string
        conversionRate:
          type: string
        convertedAmount:
          type: string
        imageUrl:
          type: string
        claimMonth:
          type: string
          nullable: true
          description: "YYYY-MM override for which claim period this receipt belongs to"
    ReceiptResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyId:
          type: integer
        categoryId:
          type: integer
          nullable: true
        claimPeriodId:
          type: integer
          nullable: true
        description:
          type: string
        receiptDate:
          type: string
          format: date
        currency:
          type: string
        amount:
          type: string
        conversionRate:
          type: string
        convertedAmount:
          type: string
        imageUrl:
          type: string
          nullable: true
        claimMonth:
          type: string
          nullable: true
          description: "YYYY-MM override for which claim period this receipt belongs to"
        categoryName:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - companyId
        - description
        - receiptDate
        - currency
        - amount
        - conversionRate
        - convertedAmount
    ParseReceiptImageBody:
      type: object
      properties:
        imageUrl:
          type: string
      required:
        - imageUrl
    ParsedReceiptData:
      type: object
      properties:
        amount:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        receiptType:
          type: string
          enum: [standard, transfer]
        confidence:
          type: number
      required:
        - confidence
        - receiptType
    ExchangeRateResponse:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        rate:
          type: number
        source:
          type: string
        timestamp:
          type: string
      required:
        - from
        - to
        - rate
        - source
...
  securitySchemes:
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram Web App initData string (HMAC-verified, 1-hour TTL)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT session token from Telegram Login Widget auth
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Server-to-server authentication key (used by Telegram bot only)